# Simplify isAlive / checkValidState methods for implicit scope