@mcimadamore 2021-04-26
Can I please ask to pause micro-optimizations for the time being? I'm glad that we're looking for opportunities to improve the code, but I think for now we should focus on stability ahead of integration into mainline JDK. Thanks!