@mcimadamore 2021-04-26
Also: ImplicitScopeImpl derives from SharedScopeImpl, whose checkValidState is as follows:

```
@Override
    public void checkValidState() {
        if (state < ALIVE) {
            throw ScopedAccessError.INSTANCE;
        }
    }
 ```
 
 Note that this just uses a simple plain access (which is safe because of the VM hackery we do to support shared memory access).